Collapse Horizontal Movement into Single UInputAction
Previously, we relied on 2 UInputActions to map the movement keys, A and D, to left and right movement, respectively.

This change consolidates the mapping into a single UInputAction and handles direction of movement by querying the UEnhancedInputLocalPlayerSubsystem's mapped keys to check which are depressed, according to the Players' Player Controller.